moves dao/model for customer module
